+++ src/Settler.sol
@@ -60,6 +60,8 @@ abstract contract Settler is ISettlerTakerSubmitted, Permit2PaymentTakerSubmitte
                     revert(0x1c, 0x44)
                 }
             }
+        } else if (action == uint32(ISettlerActions.CHECK_SLIPPAGE.selector)) {
+            _checkSlippageAndTransfer(slippage, abi.decode(data, (bool)));
         } else {
             return false;
         }