+++ CHANGELOG.md
@@ -10,13 +10,19 @@
   * Gnosis (XDAI)
 * Remove `MAVERICKV2_VIP` and `METATXN_MAVERICKV2_VIP` actions from Arbitrum,
   Base, Bnb, Mainnet, Scroll, and Sepolia. The `MAVERICKV2` action remains.
+* Update `RENEGADE` signature: add `baseForQuote` direction flag and `minBuyAmount` slippage protection
 
 ### Non-breaking changes
 
 * Add Thena UniV3 Algebra style fork to Bnb with fork ID 37
 * EkuboV2 VIP actions `EKUBO_VIP` and `METATXN_EKUBO_VIP` are not
   supported anymore and were removed from `ISettlerActions.sol`.
-* Update Arbitrum to the Osaka EVM hardfork
+* Add new `CHECK_SLIPPAGE` action to taker-submitted Settlers
+* Update the following chains to the Osaka EVM hardfork:
+  * Arbitrum
+  * Polygon
+  * Monad
+  * Scroll
 
 ## 2026-03-17